Create foo5.py
Write a similar script to foo1.py but this time use a simpler observation model where the unknown parameter is a linear transformation of the measurement. 

if theta on line 13 is switched between 1 and 3 then this will actually learn the correct parameter of the observation system. obviously it would be nicer to do this with a more interesting parameter dependence. also the system parameter alpha is assumed known. the next steo will be to simultaneously estimate both alpha and theta.